refactor: dotenv, envalid & graphql_request
