+++ env.ts
@@ -1,0 +1,8 @@
+import { config } from "dotenv";
+import { cleanEnv, str } from "envalid";
+
+await config({ export: true });
+
+export default cleanEnv(Deno.env.toObject(), {
+  BOT_TOKEN: str(),
+});
