+++ import_map.json
@@ -2,8 +2,10 @@
   "imports": {
     "grammy": "https://deno.land/x/grammy@v1.11.0/mod.ts",
     "grammy/": "https://deno.land/x/grammy@v1.11.0/",
-    "gql": "https://deno.land/x/graphql_request@v4.1.0/mod.ts",
-    "server": "https://deno.land/std@0.153.0/http/server.ts",
-    "testing/": "https://deno.land/std@0.153.0/testing/"
+    "gql": "https://esm.sh/graphql-request@5.0.0",
+    "dotenv": "https://deno.land/std@0.154.0/dotenv/mod.ts",
+    "envalid": "https://deno.land/x/envalid@0.1.2/mod.ts",
+    "server": "https://deno.land/std@0.154.0/http/server.ts",
+    "testing/": "https://deno.land/std@0.154.0/testing/"
   }
 }
