+++ src/bot.ts
@@ -2,13 +2,9 @@ import { Bot } from "grammy";
 import Middleware from "./middleware/mod.ts";
 import Commands from "./commands/mod.ts";
 import Handlers from "./handlers/mod.ts";
+import env from "../env.ts";
 
-const token = Deno.env.get("BOT_TOKEN");
-if (!token) {
-  throw new Error("BOT_TOKEN is not defined");
-}
-
-const bot = new Bot(token);
+const bot = new Bot(env.BOT_TOKEN);
 
 bot.use(Middleware, Commands, Handlers);
